DEF-1531: Fixed a couple of issues that was discovered during the code review.

+++ engine/facebook/src/java/com/dynamo/android/facebook/FacebookJNI.java
@@ -270,7 +270,7 @@ class FacebookJNI {
 
     }
 
-    public void postEvent(String event, long valueToSum, String[] keys, String[] values) {
+    public void postEvent(String event, double valueToSum, String[] keys, String[] values) {
         facebook.postEvent(event, valueToSum, keys, values);
     }
